moved email link into a spearate component
