+++ src/pages/About.js
@@ -6,6 +6,7 @@ import raw from 'raw.macro';
 
 import Main from '../layouts/Main';
 
+// uses babel to load contents of file
 const markdown = raw('../data/about.md');
 
 const count = markdown.split(/\s+/)
